requirements.txt
ultralytics
fastapi
python-multipart
uvicorn
pillow
opencv-python-headless
pyzbar